Sun, May 21, 2023, 5:13 AM -04:00

+++ filter-less/helpers.c
@@ -1,9 +1,10 @@
 #include "helpers.h"
+#include <math.h>
 
 // Convert image to grayscale
 void grayscale(int height, int width, RGBTRIPLE image[height][width])
 {
-    long int average = image[height][width].rgbtRed + image[height][width].rgbtGreen + image[height][width].rgbtBlue / 3.0;
+    int average = image[height][width].rgbtRed + image[height][width].rgbtGreen + image[height][width].rgbtBlue / 3.0;
 
     for (int i = 0; i < height; i++)
     {
